requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
cryptography==37.0.4
cycler==0.11.0
distlib==0.3.6
Django==4.1.3
djangorestframework==3.14.0
filelock==3.8.0
fonttools==4.33.3
h11==0.13.0
html5lib==1.1
identify==2.5.9
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
nodeenv==1.7.0
numpy==1.22.4
outcome==1.2.0
packaging==21.3
Pillow==9.1.1
platformdirs==2.5.4
pre-commit==2.20.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PySocks==1.7.1
python-dateutil==2.8.2
pytube==11.0.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
speedtest-cli==2.1.3
sqlparse==0.4.3
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.6
urllib3==1.26.10
virtualenv==20.16.7
webencodings==0.5.1
wsproto==1.1.0
